tests/requirements-test.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
numpy>=1.21.0
faiss-cpu>=1.7.0
elasticsearch>=8.0.0
torch>=2.0.0
transformers>=4.30.0
mock>=5.0.0 